Check that there are releases.
Previously if there were no releases the user would get an error that
references must be not be empty. It took me awhile to realize that an
empty reference there was because there was no release. I incorrectly
assumed that tags = releases. Thats not the case. While tags show up as
releases, you have to create an actual release for getLatestRelease to
work.

Also fix some !! truthy/falsey comparsions.