Change the order of logs to match git log.
Many users are surprised that the default is to reverse the order of
logs. This is not the command line default. At the time I first wrote
this it made sense to me but I do not remember why. So its a breaking
change so we will rev the major. Users that want to keep the old way
will need to add `reverse: true` to their workflow.